move selianize to its own package
