+++ package.json
@@ -19,7 +19,7 @@
   },
   "jest": {
     "testMatch": [
-      "**/packages/**/src/**/__test?(s)__/**/*.js?(x)"
+      "**/packages/**/__test?(s)__/**/*.js?(x)"
     ]
   },
   "devDependencies": {
