+++ packages/selianize/.babelrc
@@ -1,0 +1,3 @@
+{
+  presets: [ "node6" ]
+}
