+++ packages/selianize/__tests__/configuration.spec.js
@@ -15,10 +15,13 @@
 // specific language governing permissions and limitations
 // under the License.
 
-import template from "../../selianize/template";
+import ConfigurationEmitter from "../src/configuration";
 
-describe("Code template", () => {
-  it("should return the export header", () => {
-    expect(template.bootstrap()).toBe("// This file was generated using Selenium IDE\nconst By = require('selenium-webdriver').By;const until = require('selenium-webdriver').until;jest.setTimeout(30000);afterAll(() => {Runner.cleaup();});");
+describe("configuration code emitter", () => {
+  it("should emit project configuration", () => {
+    const project = {
+      url: "http://www.seleniumhq.org"
+    };
+    expect(ConfigurationEmitter.emit(project)).toBe(`const BASE_URL = '${project.url}';`);
   });
 });
