+++ packages/selianize/src/command.js
