+++ packages/selianize/src/configuration.js
