+++ packages/selianize/src/selection.js
